[refactor] Note와 Notepage 부분 Component 분리
