+++ src/components/Note/BottomNote.tsx
@@ -1,0 +1,5 @@
+import { BottomNoteContainer } from "@/components/StyleComponent/NoteStyle";
+
+export function BottomNote() {
+  return <BottomNoteContainer>I'm BottomNote</BottomNoteContainer>;
+}
