+++ src/components/Note/CommentsPage.tsx
@@ -1,0 +1,5 @@
+import { CommentsContainer } from "../StyleComponent/NoteStyle";
+
+export function CommentsPage() {
+  return <CommentsContainer>I'm Comments container</CommentsContainer>;
+}
