+++ src/components/Note/MiddleNote.tsx
@@ -1,0 +1,12 @@
+import {
+  MiddleNoteContainer,
+  NoteTitle,
+} from "@/components/StyleComponent/NoteStyle";
+
+export function MiddleNote() {
+  return (
+    <MiddleNoteContainer>
+      <NoteTitle>전체노트</NoteTitle>
+    </MiddleNoteContainer>
+  );
+}
