+++ src/pages/Home/Switch/NotePage.tsx
@@ -1,19 +1,13 @@
+import { CommentsPage } from "@/components/Note/CommentsPage";
+import { MiddleNote } from "@/components/Note/MiddleNote";
+import { TextPage } from "@/components/Note/TextPage";
+import { TopNote } from "@/components/Note/TopNote";
 import {
   CommentsContainer,
   TextContainer,
 } from "@/components/StyleComponent/NoteStyle";
 import { Outlet, useParams } from "react-router-dom";
 import styled from "styled-components";
-import { MiddleNote, TopNote } from "./Note";
-
-function TextPage() {
-  const { noteId } = useParams();
-  return <TextContainer>i'm textContainer</TextContainer>;
-}
-
-function CommentsPage() {
-  return <CommentsContainer>I'm Comments container</CommentsContainer>;
-}
 
 function NotePage() {
   return (
